+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # CHANGELOG
 
+* `2.0.81` Release - [2.0.81](#2081)
 * `2.0.80` Release - [2.0.80](#2080)
 * `2.0.79` Release - [2.0.79](#2079)
 * `2.0.78` Release - [2.0.78](#2078)
@@ -8,6 +9,14 @@
 * `0.72.x` Releases - [0.72.0](#0720)
 * `0.2.x` Releases - [0.2.0](#020) | [0.2.1](#021) | [0.2.2](#022)
 * `0.1.x` Releases - [0.1.0](#010) | [0.1.1](#011) | [0.1.2](#012) | [0.1.3](#013) | [0.1.4](#014) | [0.1.5](#015) | [0.1.6](#016) | [0.1.7](#017) | [0.1.8](#018) | [0.1.9](#019)
+
+## 2.0.81
+
+#### Features
+* `EMP-11171` `Exposure` now supports *tvOS*.
+* Added `resonseData` returning raw `Data` to `ExposureRequest`
+* Added `ExposureApi` integration point for querying a generic *Exposure* endpoint.
+* Added `LastViewedList` struct to represent *userPlayHistory* endpoint.
 
 ## 2.0.80